Delete unused objects and move Session into Schedule
